Make common cli flags as "globals"
For example: `--quiet` can be accesible at `wallust -q` and
`wallust run -q`.

Globals allow to group common arguments into the "global" space, which
is the one without a subcommand. Propagation of such cli flag is done
for ease of use (and be somewhat compatible with older behaviour):
`wallust -q run` is the same as `wallust run -q`