colorspaces: implement lchansi colors
This required a whole amount of re design (and still is a WIP), since
a type can't have 2 trait implementations, but LCHANSI required to
modify "gather_cols".

The solution to this was to a PhantomData type to ColorHisto, similar
approach taken by the palette crate, in order to mark some versions of
colorspaces.

This, howevers, requires some more cleanup and LCHANSI as well.

TODO the way sorting and deduping works after gathering the colors
should be reworked, since LCHANSI doesn't need that.