template: fix up compiler errors and add a loader for minijinja
For some reason, the compiler is throwing compiler errors about the need
to specify generic types on `.into_format()`. Problably some internal
change in palette.

Given this <https://github.com/mitsuhiko/minijinja/discussions/509#discussioncomment-9301834> discussion I learned that I need a loader in order to enable the use of `import` or `from` or `extends` and other keywords that involve other files (or manually add `.add_template()`, but looks like that requires to specify the whole path).