backends: wal: fix #55 by reading literal u8 rgb values
This could help performance, since there is nothing to interpret (other
than converting u8s to strings).

note: This is related to #51, and I thought convert used sRGB. Sometimes
it uses sRGBA as well, hexadecimally talking. With this method, it
should work in either case, since we take the first 3 values only.