colorspaces: improve filter
In some cases this doesn't modify much the result palette. When it does,
it improves incredibly the scheme, since it has much more precision on
what chroma value to use (on LCH) or brightness (on all colorspaces).

Performace: This slows downs with the `Full` backend, since it makes
operations directly on the values gathered.